refactor: adjusted the type structure
