+++ components/admin/copyright/CopyrightEditSheet.tsx
@@ -3,7 +3,8 @@
 import { Sheet, SheetContent, SheetHeader, SheetTitle } from '~/components/ui/sheet'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '~/components/ui/select'
 import { useButtonStore } from '~/app/providers/button-store-Providers'
-import { HandleProps, CopyrightType } from '~/types'
+import type { CopyrightType } from '~/types'
+import type { HandleProps } from '~/types/props'
 import { useSWRHydrated } from '~/hooks/useSWRHydrated'
 import React, { useState } from 'react'
 import { toast } from 'sonner'
