+++ hooks/useSWRPageTotalHook.ts
@@ -1,5 +1,5 @@
 import useSWR from 'swr'
-import { ImageHandleProps } from '~/types'
+import type { ImageHandleProps } from '~/types/props'
 
 export const useSWRPageTotalHook = ({ args, totalHandle, album }: ImageHandleProps)   => {
   const { data, error, isLoading, isValidating, mutate } = useSWR([args, album],
