+++ types/index.ts
@@ -1,34 +1,4 @@
-export type AlbumDataProps = {
-  data: AlbumType[]
-  customFoldAlbumEnable: boolean
-  customFoldAlbumCount: number
-}
-
-export type HandleProps = {
-  handle: () => any
-  args: string
-}
-
-export type ImageServerHandleProps = {
-  handle: (pageNum: number, tag: string) => any
-  args: string
-  totalHandle: (tag: string) => any
-}
-
-export type ImageHandleProps = {
-  handle: (pageNum: number, tag: string) => any
-  args: string
-  album: string
-  totalHandle: (tag: string) => any
-  configHandle: () => any
-  randomShow?: boolean
-}
-
-export type LinkProps = {
-  handle: () => any
-  args: string
-  data: any
-}
+// 数据库表结构类型
 
 export type AlbumType = {
   id: string;
@@ -106,36 +76,4 @@ export type Config = {
   config_key: string;
   config_value: string;
   detail: string;
-}
-
-export type AlbumListProps = {
-  data: AlbumType[]
-}
-
-export type ImageDataProps = {
-  data: ImageType
-}
-
-export type ImageListDataProps = {
-  data: ImageType[]
-}
-
-export type AnalysisDataProps = {
-  data: {
-    total: number;
-    showTotal: number;
-    crTotal: number;
-    tagsTotal: number;
-    cameraStats: Array<{
-      camera: string;
-      lens: string;
-      count: number;
-    }>;
-    result: Array<{
-      name: string;
-      value: string;
-      total: number;
-      show_total: number;
-    }>;
-  }
 }
